feat: cleanup utils
