+++ src/utils/post-data.ts
@@ -1,17 +1,5 @@
 import { Price } from '~/types/types';
 
-export const getURL = () => {
-  let url =
-    process.env.NEXT_PUBLIC_SITE_URL ??
-    process.env.NEXT_PUBLIC_VERCEL_URL ??
-    'http://localhost:3000/';
-
-  url = url.includes('http') ? url : `https://${url}`;
-  url = url.charAt(url.length - 1) === '/' ? url : `${url}/`;
-
-  return url;
-};
-
 export const postData = async ({
   url,
   data,
@@ -35,10 +23,4 @@ export const postData = async ({
   }
 
   return res.json();
-};
-
-export const toDateTime = (secs: number) => {
-  var t = new Date('1970-01-01T00:30:00Z');
-  t.setSeconds(secs);
-  return t;
 };
